+++ cutadapt/__init__.py
@@ -1,4 +1,5 @@
-from __future__ import print_function
+# coding: utf-8
+from __future__ import print_function, division, absolute_import
 import sys
 
 __version__ = '1.8.dev0'