+++ tests/testcolorspace.py
@@ -1,3 +1,4 @@
+# coding: utf-8
 from __future__ import print_function, division, absolute_import
 
 from cutadapt.colorspace import encode, decode